Time: 44 ms (32.72%), Space: 100.7 MB (35.29%) - LeetHub

+++ 3430-count-days-without-meetings/3430-count-days-without-meetings.java
@@ -11,7 +11,7 @@ class Solution {
             totalMeetings += temp>0 ? temp: 0;
             lastDay = Math.max(meetings[i][1]+1,lastDay);
         }
-        return (int) (days-totalMeetings); 
+        return days-totalMeetings; 
 
 
     }
